// src/app/core/interfaces/product/productsResponse.ts
// Generated by https://quicktype.io

export interface ProductsResponse {
    totalElements:    number;
    totalPages:       number;
    size:             number;
    content:         Product[];
    number:           number;
    sort:             Sort[];
    numberOfElements: number;
    first:            boolean;
    last:             boolean;
    pageable:         Pageable;
    empty:            boolean;
}

export interface Product {
    skuCode:     string;
    name:        string;
    description: string;
    category:    string;
    brand:       string;
    price:       number;
    sizes:       string[];
    colors:      string[];
    material:    string;
    isActive:    boolean;
}

export interface Pageable {
    offset:     number;
    sort:       Sort[];
    paged:      boolean;
    unpaged:    boolean;
    pageNumber: number;
    pageSize:   number;
}

export interface Sort {
    direction:    string;
    nullHandling: string;
    ascending:    boolean;
    property:     string;
    ignoreCase:   boolean;
}